created a components folder in the app's directory and moved salary and attendance files inside
