+++ app/components/salary.tsx
